운동량 데이터 전송 수정

+++ main/hx711_lib.c
@@ -51,36 +51,37 @@ static portMUX_TYPE mux = portMUX_INITIALIZER_UNLOCKED;
 #define BIT64 BIT
 #endif
 
-static uint32_t read_raw(gpio_num_t dout, gpio_num_t pd_sck, hx711_gain_t gain)
+static int32_t read_raw(gpio_num_t dout, gpio_num_t pd_sck, hx711_gain_t gain)
 {
-
     portENTER_CRITICAL(&mux);
 
-    // read data
     uint32_t data = 0;
+
     for (size_t i = 0; i < 24; i++)
     {
         gpio_set_level(pd_sck, 1);
         esp_rom_delay_us(1);
+
         data |= gpio_get_level(dout) << (23 - i);
+
         gpio_set_level(pd_sck, 0);
         esp_rom_delay_us(1);
     }
 
-    // config gain + channel for next read
+    // gain/channel select
     for (size_t i = 0; i <= gain; i++)
     {
         gpio_set_level(pd_sck, 1);
         esp_rom_delay_us(1);
+
         gpio_set_level(pd_sck, 0);
         esp_rom_delay_us(1);
     }
 
     portEXIT_CRITICAL(&mux);
 
-    return data;
+    return (int32_t)data;
 }
-
 ///////////////////////////////////////////////////////////////////////////////
 
 esp_err_t hx711_init(hx711_t *dev)
@@ -153,7 +154,7 @@ esp_err_t hx711_read_data(hx711_t *dev, int32_t *data)
 {
     CHECK_ARG(dev && data);
 
-    uint32_t raw = read_raw(dev->dout, dev->pd_sck, dev->gain);
+    int32_t raw = read_raw(dev->dout, dev->pd_sck, dev->gain);
     if (raw & 0x800000)
         raw |= 0xff000000;
     *data = *((int32_t *)&raw);
